Updates problem details code files
