+++ src/Client/HealthDataClientException.cs
@@ -1,32 +1,8 @@
 ﻿namespace MyBp.Client
 {
-    using System;
     using Models;
     using Microsoft.AspNetCore.Mvc;
-
-    public class ProblemDetailsException : ProblemDetailsException<ProblemDetails>
-    {
-        public ProblemDetailsException(ProblemDetails problemDetails) : base(problemDetails)
-        {
-        }
-    }
-
-    public class ProblemDetailsException<T> : ApplicationException where T : ProblemDetails
-    {
-        public T ProblemDetails { get; }
-        
-        public ProblemDetailsException(T problemDetails) : base(problemDetails.Detail)
-        {
-            this.ProblemDetails = problemDetails;
-        }
-    }
-
-    public class HealthDataClientValidationException : ProblemDetailsException<ValidationProblemDetails>
-    {
-        public HealthDataClientValidationException(ValidationProblemDetails problemDetails) : base(problemDetails)
-        {
-        }
-    }
+    using MyBp.Exceptions;
 
     public class HealthDataClientException : ProblemDetailsException<ProblemDetails>
     {
