+++ src/Client/HealthDataClientValidationException.cs
@@ -1,0 +1,12 @@
+namespace MyBp.Client
+{
+    using Microsoft.AspNetCore.Mvc;
+    using MyBp.Exceptions;
+
+    public class HealthDataClientValidationException : ProblemDetailsException<ValidationProblemDetails>
+    {
+        public HealthDataClientValidationException(ValidationProblemDetails problemDetails) : base(problemDetails)
+        {
+        }
+    }
+}
