+++ src/Views/_ViewImports.cshtml
@@ -1,4 +1,5 @@
 @using MyBp
 @using MyBp.Models
+@using Microsoft.AspNetCore.Mvc
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 @addTagHelper *, MyBp
